refactor(remote): extract stderr redaction into orchestrator submodule
Move the azlin/Azure stderr redaction helpers (GUID/SAS scrubbing) and
their tests into orchestrator/redact.rs so orchestrator.rs returns to
<=500 lines, satisfying the remote_rust_modules_stay_under_500_lines
guard (issue #536) that the added redaction feature pushed over budget.
Non-breaking: redact_sensitive is used identically at both cleanup call
sites; behavior is unchanged.
